+++ app/build.gradle
@@ -7,11 +7,11 @@ plugins {
 }
 
 android {
-    namespace 'tech.realycorp.letro'
+    namespace 'tech.relaycorp.letro'
     compileSdk 34
 
     defaultConfig {
-        applicationId "tech.realycorp.letro"
+        applicationId "tech.relaycorp.letro"
         minSdk 23
         targetSdk 34
         versionCode 1
