--- app/src/main/java/tech/realycorp/letro/App.kt
@@ -1,7 +1,0 @@
-package tech.realycorp.letro
-
-import android.app.Application
-import dagger.hilt.android.HiltAndroidApp
-
-@HiltAndroidApp
-open class App : Application()
