--- app/src/main/java/tech/realycorp/letro/data/GatewayAvailabilityDataModel.kt
@@ -1,7 +1,0 @@
-package tech.realycorp.letro.data
-
-sealed interface GatewayAvailabilityDataModel {
-    object Unavailable : GatewayAvailabilityDataModel
-    object Available : GatewayAvailabilityDataModel
-    object Unknown : GatewayAvailabilityDataModel
-}
