--- app/src/main/java/tech/realycorp/letro/data/UserDataModel.kt
@@ -1,6 +1,0 @@
-package tech.realycorp.letro.data
-
-data class UserDataModel(
-    val username: String,
-    val connections: List<String> = emptyList(),
-)
