--- app/src/main/java/tech/realycorp/letro/ui/main/FirstNavigationUIModel.kt
@@ -1,7 +1,0 @@
-package tech.realycorp.letro.ui.main
-
-sealed interface FirstNavigationUIModel {
-    object NoGateway : FirstNavigationUIModel
-    object AccountCreation : FirstNavigationUIModel
-    object Splash : FirstNavigationUIModel
-}
