--- app/src/main/java/tech/realycorp/letro/ui/onboarding/accountCreation/AccountCreationUIState.kt
@@ -1,5 +1,0 @@
-package tech.realycorp.letro.ui.onboarding.accountCreation
-
-data class AccountCreationUIState(
-    val username: String = "",
-)
