--- app/src/main/java/tech/realycorp/letro/ui/theme/Color.kt
@@ -1,41 +1,0 @@
-package tech.realycorp.letro.ui.theme
-
-import androidx.compose.ui.graphics.Color
-
-val Purple80 = Color(0xFFD0BCFF)
-val PurpleGrey80 = Color(0xFFCCC2DC)
-val Pink80 = Color(0xFFEFB8C8)
-
-val Purple40 = Color(0xFF6650a4)
-val PurpleGrey40 = Color(0xFF625b71)
-val Pink40 = Color(0xFF7D5260)
-
-val Grey10 = Color(0xFFFFFFFF)
-val Grey50 = Color(0xFFEAEDF6)
-val Grey60 = Color(0xFFEAEDF6)
-val Grey70 = Color(0xFFD9E0FF)
-val Grey80 = Color(0xFFADB9DC)
-val Grey90 = Color(0xFF66759D)
-val Grey100 = Color(0xFF11245A)
-
-val PrimaryMain = Color(0xFF7158E2)
-val PrimaryPurBlue = Color(0xFF4C28FF)
-val SecondarySkyBlue = Color(0xFF54AAFF)
-val SecondaryLightBlue = Color(0xFFD1E9FF)
-val SecondaryPink = Color(0xFFEF78A4)
-val SecondaryBabyPink = Color(0xFFFBDDE8)
-
-val WarningMain = Color(0xFFFEBE2B)
-val WarningBorder = Color(0xFFFFDD65)
-val WarningBackground = Color(0xFFFFF3C7)
-
-val SuccessMain = Color(0xFF1ABF6C)
-val SuccessBorder = Color(0xFF83ECB7)
-val SuccessBackground = Color(0xFFC9FFE4)
-
-val ErrorMain = Color(0xFFFF0000)
-val ErrorBorder = Color(0xFFFFB5B5)
-val ErrorBackground = Color(0xFFFFDFDF)
-
-val InfoMain = Color(0xFF3987FD)
-val InfoBackground = Color(0xFFE7F0FF)
