--- app/src/main/java/tech/realycorp/letro/ui/theme/Dimension.kt
@@ -1,12 +1,0 @@
-package tech.realycorp.letro.ui.theme
-
-import androidx.compose.ui.unit.dp
-
-val HorizontalScreenPadding = 16.dp
-val VerticalScreenPadding = 16.dp
-
-val LargePadding = 32.dp
-val ItemPadding = 8.dp
-
-val TextFieldCornerRadius = 4.dp
-val BoxCornerRadius = 8.dp
