+++ app/src/main/java/tech/relaycorp/letro/App.kt
@@ -1,0 +1,7 @@
+package tech.relaycorp.letro
+
+import android.app.Application
+import dagger.hilt.android.HiltAndroidApp
+
+@HiltAndroidApp
+open class App : Application()
