+++ app/src/main/java/tech/relaycorp/letro/data/GatewayAvailabilityDataModel.kt
@@ -1,0 +1,7 @@
+package tech.relaycorp.letro.data
+
+sealed interface GatewayAvailabilityDataModel {
+    object Unavailable : GatewayAvailabilityDataModel
+    object Available : GatewayAvailabilityDataModel
+    object Unknown : GatewayAvailabilityDataModel
+}
