+++ app/src/main/java/tech/relaycorp/letro/data/UserDataModel.kt
@@ -1,0 +1,6 @@
+package tech.relaycorp.letro.data
+
+data class UserDataModel(
+    val username: String,
+    val connections: List<String> = emptyList(),
+)
