+++ app/src/main/java/tech/relaycorp/letro/ui/main/FirstNavigationUIModel.kt
@@ -1,0 +1,7 @@
+package tech.relaycorp.letro.ui.main
+
+sealed interface FirstNavigationUIModel {
+    object NoGateway : FirstNavigationUIModel
+    object AccountCreation : FirstNavigationUIModel
+    object Splash : FirstNavigationUIModel
+}
