+++ app/src/main/java/tech/relaycorp/letro/ui/onboarding/accountCreation/AccountCreationUIState.kt
@@ -1,0 +1,5 @@
+package tech.relaycorp.letro.ui.onboarding.accountCreation
+
+data class AccountCreationUIState(
+    val username: String = "",
+)
