+++ app/src/main/java/tech/relaycorp/letro/ui/theme/Dimension.kt
@@ -1,0 +1,12 @@
+package tech.relaycorp.letro.ui.theme
+
+import androidx.compose.ui.unit.dp
+
+val HorizontalScreenPadding = 16.dp
+val VerticalScreenPadding = 16.dp
+
+val LargePadding = 32.dp
+val ItemPadding = 8.dp
+
+val TextFieldCornerRadius = 4.dp
+val BoxCornerRadius = 8.dp
